Reorganize socket-based transports
Move the socket-based transports, raw and tls, under a transport::socket
module.  This includes the frame module, as it is used by both of these
modules.
